incident-reporter/requirements.txt
opensearch-py>=2.4.0
kafka-python==2.0.2
anthropic>=0.40.0
pytest==8.2.2
pytest-mock==3.14.0
